+++ simmetrics-core/src/main/java/org/simmetrics/metrics/Levenshtein.java
@@ -21,27 +21,56 @@
  */
 package org.simmetrics.metrics;
 
+import static com.google.common.base.Preconditions.checkArgument;
 import static java.lang.Math.max;
-import static org.simmetrics.utils.Math.min3;
+import static org.simmetrics.utils.Math.min;
 
-import java.util.Objects;
-
+import org.simmetrics.StringDistance;
 import org.simmetrics.StringMetric;
 
 /**
  * Levenshtein algorithm providing a similarity measure between two strings.
- * 
- * 
- *
+ * <p>
+ * Insert/delete and substitute operations can be weighted. When the cost for
+ * substitution is zero Levenshtein does not satisfy the coincidence property.
  * <p>
  * This class is immutable and thread-safe.
  * 
  * @see <a href=" http://en.wikipedia.org/wiki/Levenshtein_distance">Wikipedia -
  *      Levenshtein distance</a>
- * 
+ * @see DamerauLevenshtein
  * 
  */
-public class Levenshtein implements StringMetric {
+public class Levenshtein implements StringMetric, StringDistance {
+
+	private final float maxCost;
+	private final float insertDelete;
+	private final float substitute;
+
+	/**
+	 * Constructs a new weighted Levenshtein metric. When the cost for
+	 * substitution is zero Levenshtein does not satisfy the coincidence
+	 * property.
+	 * 
+	 * @param insertDelete
+	 *            positive non-zero cost of an insert or deletion operation
+	 * @param substitute
+	 *            positive cost of a substitute operation
+	 */
+	public Levenshtein(float insertDelete, float substitute) {
+		checkArgument(insertDelete > 0);
+		checkArgument(substitute >= 0);
+		this.maxCost = max(insertDelete, substitute);
+		this.insertDelete = insertDelete;
+		this.substitute = substitute;
+	}
+
+	/**
+	 * Constructs a new Levenshtein metric.
+	 */
+	public Levenshtein() {
+		this(1.0f, 1.0f);
+	}
 
 	@Override
 	public float compare(final String a, final String b) {
@@ -49,56 +78,60 @@ public class Levenshtein implements StringMetric {
 			return 1.0f;
 		}
 
-		if (a.isEmpty() || b.isEmpty()) {
-			return 0.0f;
-		}
-
-		return 1.0f - (levenstein(a, b) / max(a.length(), b.length()));
+		return 1.0f - (distance(a, b) / (maxCost * max(a.length(), b.length())));
 	}
 
-	private static float levenstein(final String s, final String t) {
+	@Override
+	public float distance(final String s, final String t) {
 
-		if (Objects.equals(s, t))
-			return 0;
 		if (s.isEmpty())
 			return t.length();
 		if (t.isEmpty())
 			return s.length();
+		if (s.equals(t))
+			return 0;
 
 		final int tLength = t.length();
 		final int sLength = s.length();
-		
-		final float[] v0 = new float[tLength + 1];
-		final float[] v1 = new float[tLength + 1];
+
+		float[] swap;
+		float[] v0 = new float[tLength + 1];
+		float[] v1 = new float[tLength + 1];
 
 		// initialize v0 (the previous row of distances)
 		// this row is A[0][i]: edit distance for an empty s
 		// the distance is just the number of characters to delete from t
 		for (int i = 0; i < v0.length; i++) {
-			v0[i] = i;
+			v0[i] = i * insertDelete;
 		}
 
-		for (int i = 0; i < sLength ; i++) {
+		for (int i = 0; i < sLength; i++) {
 
 			// first element of v1 is A[i+1][0]
 			// edit distance is delete (i+1) chars from s to match empty t
-			v1[0] = i + 1;
+			v1[0] = (i + 1) * insertDelete;
 
 			for (int j = 0; j < tLength; j++) {
-				v1[j + 1] = min3(v1[j] + 1, v0[j + 1] + 1, v0[j]
-						+ (s.charAt(i) == t.charAt(j) ? 0.0f : 1.0f));
+				v1[j + 1] = min(v1[j] + insertDelete,
+						v0[j + 1] + insertDelete,
+						v0[j]
+								+ (s.charAt(i) == t.charAt(j) ? 0.0f
+										: substitute));
 			}
 
-			for (int j = 0; j < v0.length; j++) {
-				v0[j] = v1[j];
-			}
+			swap = v0;
+			v0 = v1;
+			v1 = swap;
 		}
 
-		return v1[tLength];
+		// latest results was in v1 which was swapped with v0
+		return v0[tLength];
 	}
 
 	@Override
 	public String toString() {
-		return "Levenshtein";
+		return "Levenshtein [insertDelete=" + insertDelete + ", substitute="
+				+ substitute + "]";
 	}
+
 }