+++ simmetrics-core/src/test/java/org/simmetrics/metrics/BlockDistanceTest.java
@@ -1,5 +1,5 @@
 /*
- * SimMetrics - SimMetrics is a java library of Similarity or Distance Metrics,
+ * SimMetrics SimMetrics is a java library of Similarity or Distance Metrics,
  * e.g. Levenshtein Distance, that provide float based similarity measures
  * between String Data. All metrics return consistent measures rather than
  * unbounded similarity scores.
@@ -21,42 +21,85 @@
  */
 package org.simmetrics.metrics;
 
-import org.junit.Test;
-import org.simmetrics.ListMetric;
-import org.simmetrics.metrics.BlockDistance;
+import static java.util.Arrays.asList;
+
+import java.util.List;
+
+import org.simmetrics.Distance;
+import org.simmetrics.ListDistanceTest;
+import org.simmetrics.ListMetricTest;
+import org.simmetrics.Metric;
 import org.simmetrics.tokenizers.QGram;
-import org.simmetrics.tokenizers.Whitespace;
+import org.simmetrics.tokenizers.Tokenizer;
 
-public class BlockDistanceTest extends ListMetricTest {
+@SuppressWarnings("javadoc")
+public final class BlockDistanceTest {
 
-	
-	@Override
-	public ListMetric<String> getMetric() {
-		return new BlockDistance<>();
+	public final static class DistanceList extends ListDistanceTest {
+
+		@Override
+		protected Distance<List<String>> getMetric() {
+			return new BlockDistance<>();
+		}
+
+		@Override
+		protected boolean satisfiesSubadditivity() {
+			return false;
+		}
+
+		@Override
+		protected T[] getListTests() {
+			return new T[] {
+					new T(0.0000f, "test string1", "test string1"),
+					new T(2.0000f, "test string1", "test string2"),
+					new T(1.0000f, "test", "test string2"),
+					new T(2.0000f, "", "test string2"),
+					new T(2.0000f, asList("test", null), asList("test","string2")),
+					new T(2.0000f, "aaa bbb ccc ddd", "aaa bbb ccc eee"),
+					new T(2.0000f, "aaa bbb", "aaa aaa"),
+					new T(1.0000f, "aaa", "aaa aaa"),
+					new T(2.0000f, "a b c d", "a b c e"),
+					new T(4.0000f, "a b c d", "a b e f") };
+		}
 	}
 
-	@Test
-	public void test1() {
-		testSimilarity(
-				getMetric(), 
-				new Whitespace(), 
-				new T(0.5000f, "test string1", "test string2"), 
-				new T(0.7500f, "aaa bbb ccc ddd", "aaa bbb ccc eee"),
-				new T(0.7500f, "a b c d", "a b c e"));
-	}
+	public final static class MetricList extends ListMetricTest {
 
-	@Test
-	public void test2() {
-		testSimilarity(
-				getMetric(),
-				new QGram(1) ,
-				new T(0.8333f, "Healed", "Sealed"),
-				new T(0.6153f, "Healed", "Healthy"),
-				new T(0.7272f, "Healed", "Heard"),
-				new T(0.6666f, "Healed", "Herded"),
-				new T(0.6000f, "Healed", "Help"),
-				new T(0.4000f, "Healed", "Sold"),
-				new T(0.6000f, "Healed", "Help")
-		);
+		@Override
+		protected boolean satisfiesSubadditivity() {
+			return false;
+		}
+
+		@Override
+		protected T[] getListTests() {
+			final Tokenizer q1 = new QGram(1);
+			return new T[] { 
+					new T(0.5000f, asList("test", null), asList("test","string2")),
+					new T(1.0000f, "test string1", "test string1"),
+					new T(0.5000f, "test string1", "test string2"),
+					new T(0.6666f, "test", "test string2"),
+					new T(0.0000f, "", "test string2"),
+					new T(0.7500f, "aaa bbb ccc ddd", "aaa bbb ccc eee"),
+					new T(0.5000f, "aaa bbb", "aaa aaa"),
+					new T(0.6666f, "aaa", "aaa aaa"),
+					new T(0.7500f, "a b c d", "a b c e"),
+					new T(0.5000f, "a b c d", "a b e f"),
+
+					new T(q1, 0.8333f, "Healed", "Sealed"),
+					new T(q1, 0.6153f, "Healed", "Healthy"),
+					new T(q1, 0.7272f, "Healed", "Heard"),
+					new T(q1, 0.6666f, "Healed", "Herded"),
+					new T(q1, 0.6000f, "Healed", "Help"),
+					new T(q1, 0.4000f, "Healed", "Sold"),
+					new T(q1, 0.6000f, "Healed", "Help")
+
+			};
+		}
+
+		@Override
+		protected Metric<List<String>> getMetric() {
+			return new BlockDistance<>();
+		}
+
 	}
 }