+++ simmetrics-core/src/test/java/org/simmetrics/metrics/SmithWatermanTest.java
@@ -21,24 +21,40 @@
 package org.simmetrics.metrics;
 
 import org.simmetrics.StringMetric;
+import org.simmetrics.StringMetricTest;
 import org.simmetrics.metrics.SmithWaterman;
 
 
+@SuppressWarnings("javadoc")
 public class SmithWatermanTest extends StringMetricTest {
-
+	
+	@Override
+	protected boolean satisfiesSubadditivity(){
+		return false;
+	}
+	
+	@Override
+	protected boolean satisfiesCoincidence() {
+		return false;
+	}
+	
 	@Override
 	protected StringMetric getMetric() {
 		return new SmithWaterman();
 	}
 
+	
 	@Override
-	protected T[] getTests() {
+	protected T[] getStringTests()  {
 		return new T[] {
+				new T(0.0000f, "", "eee"),
 				new T(0.0000f, "aaa", "eee"),
 				new T(0.0000f, "eee", "aaa"),
 				new T(0.0000f, "ddd", "aaa"),
 				new T(0.0000f, "aaa", "ddd"),
-				new T(0.9167f, "test string1", "test string2"),
+				new T(0.9166f, "test string1", "test string2"),
+				new T(1.0000f, "test", "test string2"),
+				new T(0.0000f, "", "test string2"),
 				new T(0.8000f, "aaa bbb ccc ddd", "aaa bbb ccc eee"),
 				new T(0.8571f, "a b c d", "a b c e"),
 				new T(0.8333f, "Healed", "Sealed"),